split Acorn classes by separate packages to preserve meta
Breaking Change for core (when Acorn is used as a lib)
Without this change cli and core libs share the same package, and we'll take the latest loaded package meta (like versions).
Splitting them allows us to keep cli module on version 'latest' but we still have versions for core.